requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
python-socketio[client]==5.11.2
pydantic==2.8.2
PyYAML==6.0.2
typer==0.12.3
rich==13.7.1
pytest==8.3.2
